Move materials into dedicated file like other routes
